# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "United Tractor API",
        "description": "United Tractor API documentation",
        "contact": {
            "email": "fajarshidik78@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login into account",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "tokoweb@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "tokoweb"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "email": {
                                                    "type": "string",
                                                    "example": "tokoweb@gmail.com"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "ASDAEOJEAOAJEOAEOA212e4nsifcwaie2h"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register new account",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "tokoweb@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "tokoweb"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "email": {
                                                    "type": "string",
                                                    "example": "tokoweb@gmail.com"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "ASDAEOJEAOAJEOAEOA212e4nsifcwaie2h"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "BearerToken": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer &lt;token&gt;)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}